+++ release-notes.md
@@ -1,2 +1,1 @@
-Release V1.0
-Updated Github Actions, and target framework to include net.6 for windows7 compatability.
+- Do not verify patch files to prevent System.Collections.Generic.KeyNotFoundException error